amongagents/requirements.txt
openai
anthropic
python-dotenv
matplotlib
numpy
